Bumped Rails 5.2 -> 6.0
(This is an rough and ready upgrade to get an idea of the
scope. Nothing about it is production-worthy.)

Modified config to use the classic preloader; ticket to investigate
switching to zeitwerk.

Also had to bump:
  `responders` 2.0 -> 3.0
  `authlogic` 3.8 -> 6.1.0
  `newrelic_rpm` 5.0 -> 7.0
(Note that these are not necessarily the final versions we will
use. Each of these gets its own investigaton ticket.)

Completely removed airbrake and associated configuration. There is a
ticket for doing this, and it should happen during the Rails upgrade.

Removed i18-tasks (a dev/test only gem). This gets an investigaton
ticket; don't know if we need it or not.